Resolve Glyph Reversed for U+05F1 HEBREW LIGATURE YIDDISH VAV YOD (issue #17)
Reverse VAV and YOD components, which had been in the wrong
right-to-left order, in error, in .glyphs file

sources/FrankRuhlLibre.glyphs

Then regenerated *some* of the fonts, specifically

  fonts/otf/FrankRuhlLibre-Black.otf
  fonts/otf/FrankRuhlLibre-Light.otf
  fonts/otf/FrankRuhlLibre-Medium.otf
  fonts/otf/FrankRuhlLibre-Regular.otf
  fonts/ttf/FrankRuhlLibre-Black.ttf
  fonts/ttf/FrankRuhlLibre-Light.ttf
  fonts/ttf/FrankRuhlLibre-Medium.ttf
  fonts/ttf/FrankRuhlLibre-Regular.ttf

Now, character U+05F1 HEBREW LIGATURE YIDDISH VAV YOD (ױ) should
appear correctly, i.e., with VAV subcomponent to the right (first in
right-to-left order) and YOD subcomponent to the left (second in
right-to-left order), whereas previously they were reversed.

NOTE: I generated fonts with the command

  fontmake -g FrankRuhlLibre.glyphs

However, this did NOT create ALL the different weights and file types
as had been present earlier.  I need help figuring out what's needed
to get them all generated now that the master source file has been
changed.  This repo seems to be missing build info, or I'm missing
some kind of assumed knowledge. I'm new to Google Font building.

NOTE 2: I did NOT create a backup version of file
FrankRuhlLibre.glyphs in the sources/Backup/ directory, since I was
not sure what the protocol is for doing that and also not sure what
would be the point of that when the source is maintained in a git
repo.