Resolve Glyph Reversed for U+05F1 HEBREW LIGATURE YIDDISH VAV YOD (issue #17)
Reverse VAV and YOD components, which had been in the wrong
right-to-left order, in error, in .glyphs file

  sources/FrankRuhlLibre.glyphs

Now, character U+05F1 HEBREW LIGATURE YIDDISH VAV YOD (ױ) should
appear correctly, i.e., with VAV subcomponent to the right (first in
right-to-left order) and YOD subcomponent to the left (second in
right-to-left order), whereas previously they were reversed.

Note: this was hand-tested by generating just a few font files and
viewing in FontForge, but I leave it to upstream developers to
perform the the full build based on this source change.